Merged PR 42157: Refactor sync 'from time' overrides
## What's being changed

The logic to do with setting a 'sync from time' via the CLI has been updated.

## Why it's being changed

Some time ago we added the ability to supply a time as an option on the command line when running AbandonedCart sync. This was strictly a debugging utility, but here's an example:
```
bin/magento dotdigital:sync AbandonedCart --from='2022-12-20 10:00:00'
```
However, this was never particularly useful. It set the supplied time as the 'to' time, then subtracted from that a) the configured interval e.g. 15 minutes, then b) the standard 5 minute AC interval. So it was still quite a fiddle to run AbandonedCart sync on demand and sync the cart you just dropped. Unsurprisingly, we hardly ever used it (instead relying on hacking the code to return the window we needed to catch our quote).

This PR changes the logic so that:
- if a from time is supplied, any other intervals are ignored. The time window will be 'time supplied' up until 'now'.
- An overridden from time will be in effect for all numbers in abandoned cart series (so all three emails would be queued in theory).
- An overridden from time also applies to the 'do not resend within' time limit, but in that case we measure the time limit duration _before_ the from time i.e. the from time becomes the to time.

The driver in this case is an attempt to sync abandoned carts from our remote test system, by automating the creation of quotes, then running the sync without having to wait for the interval to pass.

## How to review / test this change

- Turn off all abandoned cart exclusion rules
- Re-test abandoned cart program enrolment
- Re-test abandoned cart email series
- Confirm that the intervals specified in the config are still respected each and every time they are looked up
- Confirm that all 3 customer series are processed correctly
- Confirm that all 3 guest series are processed correctly
- Supply a time via an option per the above example, examine how this changes the intervals
- Configure a time limit via Configuration > Abandoned Cart, check that this is respected i.e. campaigns are not queued if the same contact has been emailed in the last x hours, automations are not queued if the same contact has been enrolled in the last x hours.

Related work items: #203475