Merged PR 43853: Improve extensibility of customer account dashboard
## What's being changed

We've made our system to customise pieces of the account dashboard more flexible, so that we can make changes to the navigation and the page content based on different configuration values in different modules.

## Why it's being changed

We needed a way to redirect to our custom layout based on certain conditions, and we also needed to change the tab label, based on a similar set of conditions. We wanted to override the default behaviour in this module, then in the SMS module, override either the default or the overrides from this module, again based on multiple conditions.

We were limited by:
a) `ifconfig` can only take one condition
b) `referenceBlock` cannot take an `ifconfig` arg

Because of this we have the `toHtml` overrides and also the observers to remove the unused navigation links. Unfortunately we still also need to keep the plugin because there are buttons elsewhere in the admin that link to /newsletter/manage and we have to redirect this traffic also.

## How to review / test this change

_Account UI_
- Disable the account connection
- Log in as a customer
- Confirm you see the default 'Newsletter Subscriptions' tab and page
- Enable the account connection (with preferences, addressbooks and datafields all switched off)
- Again, you should see the default
- Now, enable one of preferences, addressbooks or datafields
- You should now see the connector/customer/index page
- Go to My Account > Newsletter Subscriptions 'Edit' (you should still be redirected to connector/customer/index)

_CRUD operations_
- Switch on display of address books, data fields and marketing preferences
- Ensure you have some ABs, DFs and prefs to display (you'll need to set these as 'public' in DD)
- Marketing prefs should have some nested preferences
- Now you can thoroughly test interacting with the settings panel

Related work items: #201799, #202552, #202556