Merged PR 44532: Add check for customer email before loading from table
## What's being changed

In this PR we check if customer is logged in before we attempt to load a contact from our table, in the AccountSubscription viewModel.

## Why it's being changed

During QA we spotted an anomaly that causes UI bugs in frontend. More specifically if a contact without an email exists in email_contact table then in customer registration that contact is loaded. This could result in the consent fieldset checkbox being checked by default, and the mobile number field number being pre-populated with the no-email contact's mobile number.

## How to review / test this change

- Create a contact in email_contact table with empty string email address mobile_number and sms_subscriber_status = 1
- Go to customer registration page
- Make sure consent fieldset checkbox is not checked and behaves as expected.

## Notes
Although empty email addresses in email_contact table shouldn't be accepted (we have a bug in the backlog to address this) we need to ensure that we handle this case too in viewmodel.

Related work items: #208207