Merged PR 44748: Replace getSize for count of filtered collection in guest sync
## What's being changed

We've replaced `getSize()` with a count on the collection's items, in guest sync.

## Why it's being changed

In this context, `getSize()` returns the total collection count, not the count of the filtered collection with offset applied. So, if the first `getSize()` returned > 1, so will the `getSize()` on all subsequent loops. The impact of this is minor, it just means we run more code than we need to, instead of breaking and proceeding to the batch processor.

## How to review / test this change

- Start with at least 1 guest to be synced
- Run guest sync with a breakpoint on L189
- Ensure that we break after the first loop

Related work items: #209647