Merged PR 44886: Update unit tests to be compatible with PHP 8.2
## What's being changed

Dynamic declaration of property mocks in our unit tests is now removed. We also fixed a unit test that failed occasionally.

## Why it's being changed

The dynamic property declaration is now deprecated in PHP 8.2. The failing test was because `rand(1, 99)` when called 5 times could return 2 of the same number in a set of 5, which meant that that particular id was not added to the `$data` array.

## How to review / test this change

- Make sure that unit tests pass without errors and warnings in PHP 8.2

Related work items: #209883