Merged PR 45078: Added hidden config key to manage consent sync
## What's being changed

We have added new configuration key to govern the state of the new consent sync features.

## Why it's being changed

We needed a way to disable and enable the consent sync in case the merchant would not want consent to be synced with their contacts. The setting defaults to 1 (and is not editable in the admin) because it is assumed that merchants will want to sync consent if they are capturing it.

## How to review / test this change

1. Run ` bin/magento config:show sync_settings/sync/consent_enabled` value should be 1 by default
2. Run `bin/magento config:set sync_settings/sync/consent_enabled 0`
3. Run ` bin/magento config:show sync_settings/sync/consent_enabled` value should be 0 now

## Notes
Be sure to run the `bin/magento setup:upgrade` command first.

Related work items: #210729