Merged PR 45205: Optimise writing of csv to export files
## What's being changed

We have applied a patch provided by one of our clients to optimize the file writes of our sync csvs, by moving the file open call to outside the batch line item loop and the close file call to after the loop has pushed the data to the file.

## Why it's being changed

We were opening and closing the file for each item in the loop, which was unnecessarily resource intensive, especially on some types of infrastructure.

## How to review / test this change

1. Seed some contacts using the Utilities fixture.
2. Run the customer sync and importer to ensure the records get imported to DD correctly.
3. Check the csv generated for any invalid or anomalous records.
4. Run smoke tests on the following syncs:
-  Catalog
-  Customer
-  Guest
-  Subscriber

Related work items: #211502