Merged PR 45014: Prevent saving empty string for email
## What's being changed

This PR prevents the storing of an empty string for the `email` column in the `email_contact` table. All calls to `setEmail()` on the contact have been reviewed, most of these being indirect calls via the `loadByCustomerEmail()` method (which we've now deprecated).

## Why it's being changed

We noticed in a couple of environments that rows had been stored in `email_contact` with an empty email value, so we decided to tighten this up.

In the process of throwing the exception and needing to catch it in a various places, we've reviewed usages of the Contact model's `loadByCustomerEmail()` method. We've deprecated this method because it did two things - it loaded a row from the Contact Collection, OR it created a stub of a Contact model if it didn't find anything. This felt muddled.

We reviewed usages of this method based on the following criteria:
1. If the calling code is in an unused/deprecated method, ignore
2. If the method is already being called inside a try/catch block, it's safe so ignore
3. If we know a contact will have been created already (as in the customer account), replace method calls with the Contact collection's own `loadByCustomerEmail()` method
4. For all other cases, wrap calling code in a try/catch

The only case where we went further than this was in the Newsletter controller, where the code seemed to call for a small refactor.

## How to review / test this change

Smoke test functionality in the following areas:
- Customer account (update subscription)
- Automation sync (test processing automations for new and existing contacts)
- Remove Contact observer (delete a Magento customer or subscriber)
- Order Place After observer (place an order as a guest)
- Register observer [SMS]

Hack the code in a couple of places to `setEmail('')` and check an exception is logged.

Have we missed any places where an empty email will throw an uncaught exception and thus break something?

Related work items: #208588