Merged PR 46110: Trim credentials to avoid V3 failed authorisation
## What's being changed

We trim any spaces from ddg credentials before we authorise the V3 client directly from helper.

## Why it's being changed

Sometimes the copy paste of the api user from Dotdigital platform to Magento connector leaves empty spaces in the api username. This can lead to authorisation error and failure api requests. For the V2 client we were already using trim when setting the creds and endpoint - but we weren't with the V3 client.

## How to review / test this change

- Copy paste api user name from ddg platform to your api user name in Dotdigital > Accounts section and make sure that a space left in the end.
- Perform any V3 requests.
- Make sure no authorisation errors happening.
- Repeat the above steps and make sure that any V2 requests works as expected

Related work items: #216215