Merged PR 42650: Replace Zend with Laminas
## What's being changed

We're replacing any references to Zend classes with their Laminas equivalents.

## Why it's being changed

Magento [switched from Zend to Laminas](https://community.magento.com/t5/Magento-DevBlog/Migration-of-Zend-Framework-to-the-Laminas-Project/ba-p/443251) in 2.3.5. We have only just been able to make this change on our side because we are shortly to increase our requirements to PHP 7.4 / Magento 2.3.7+.

## How to review / test this change

- Confirm this module has no remaining references to Zend classes (classes in `magento/zend-db` are OK because they are still bundled)
- Test the chat bubble loads and chat functions as expected
- Test update profile works (i.e. if you start a chat, then register or log in as a customer, the DD chat is updated with your name)

Related work items: #147470