Merged PR 46453: Remove non-customer row if customer changes email to match existing
## What's being changed

The CreateUpdateContact observer.

## Why it's being changed

If a customer changes her email, and the new email matches an existing non-customer row in the current website scope, we had a problem. Previously our observer would update matching rows by customer id, and this wouldn't include guests or subscribers. The result was that we would have multiple contacts in our table for a given website - and, since there are quite a few places where we fetch a contact by email address as opposed to customer id, this could lead to missing or incorrect data.

The modified observer checks for orphaned rows i.e. if an email change is detected, we check to see if there are any rows matching the new email address in the table (and without a customer id), and if so delete them.

## How to review / test this change

_Removing orphaned guest_
- Place an order as guest with email chaz1@emailsim.io
- Now register a customer chaz2@emailsim.io
- In your chaz2 account, change your email address to chaz1@emailsim.io
- Confirm that you end up with 1 customer row for chaz1@emailsim.io
- Guest rows for chaz1@emailsim.io in other scopes should not be deleted
- Turn on GAS mode and set up a customer with multiple rows for different websites
- Repeat the test. You should see all matching customers updated, and any guest row for the new email address deleted.

_Removing orphaned subscriber_
- Subscribe with sub1@emailsim.io
- Now register a customer sub2@emailsim.io
- In your sub2 account, change your email address to sub1@emailsim.io
- Confirm that you end up with 1 customer row for sub1@emailsim.io

_Admin re-association_
- Turn on GAS
- Create chaz1website1@emailsim.io associated to website 1
- In the admin, change email to chaz1website2@emailsim.io and association to website 2
- Save
- Confirm you have two rows for chaz1website2@emailsim.io

Related work items: #216696