Merged PR 46688: Use a factory to fetch columns for current website
The subscriber exporter for guest subscribers with sales data.

It is possible in some syncs across multiple websites that the subscriber exporter's columns are set according to one website's configuration, but this does not match another website's configuration - leading to incorrect columns in the export.

Evidently we assumed when writing this code that a regular subscriber export for a given website would _always_ precede a subscriber with sales export - but this is not the case, there may only be a cohort for subscribers with sales data for a certain website.

- Turn on Developer > Enable Subscriber Sales Data Sync at default level
- Turn on Customers > Newsletter > Need to confirm at website 2 level (leave as No everywhere else)
- Subscribe an email address on website 1
- Place a guest order with this email address on website 1
- Subscribe an email address on website 2
- Confirm your subscription (either click on the link in the email, or if you haven't set that up, set status to 1 in newsletter_subscriber, and set 1 in email_contact.is_subscriber and email_contact.subscriber_status
- Place a guest order with this email address on website 2
- Ensure you have no other subscribers to sync besides the above 2 guest subscribers
- Run subscriber sync
- Confirm that 'OptInType' is output in the export for website 2 but not in the export for website 1

Related work items: #219151