Merged PR 46839: Set a limit for number of consent records pushed per contact
## What's being changed

We've added a limit of 5 items of consent insight that can be pushed per contact in any given bulk import via the contacts service. If a contact has more than 5 consent records to be synced, we skip any after 5 until the subsequent sync.

## Why it's being changed

This is a current limitation of the API.

## How to review / test this change

- Build up more than 5 consent record rows for a given contact
- Run consent sync
- Only the first 5 should be marked as imported
- Total synced records should reflect the number actually being synced
- Make sure to check the consent insight data being posted to DD (I have changed from using array keys to using magic getters)
- Make sure that empty consent text still throws an exception, and that this is logged with the email address

Related work items: #219157