Merged PR 46762: Try/catch data migration steps
## What's being changed

We've wrapped each migration step in a try/catch block to prevent the whole process being derailed by a single error.

## Why it's being changed

Two of our migrate steps (the ones that extend `AbstractBulkUpdater` i.e. `UpdateEmailContactTableCustomerSales` and `UpdateEmailContactTableGuestSales`) are not compatible with split database installations. This is because we need to use the sales_order.store_id to JOIN the store table in order to fetch a website_id for the query - and this use of JOIN on `sales_order` is a specific limitation of split db.

We've added the try/catch to keep surfacing the error whilst proceeding with other migration steps.

## How to review / test this change

- Set up a split db Magento
- Run bin/magento dotdigital:migrate
- See an error in the logs
- See an error output to the terminal

## Notes

- `UpdateEmailContactTableCustomerSales` only runs if Global Account Sharing is enabled
- `UpdateEmailContactTableGuestSales` will update any customers as `is_guest` if they have previously placed a guest order. This migration step is intended to maintain data parity before and after migration - but it isn't considered critical.

Related work items: #220050