Merged PR 47077: Ensure that null is not passed to trim in API helper methods
## What's being changed

The API username and password helper methods.

## Why it's being changed

A recent change to trim these values had an unexpected regression in a data patch. When this patch is run, if our module is not yet installed (i.e. it is being installed for the first time) then the relevant configs will not yet exist.

## How to review / test this change

- Go to a cold install (or observe the Magento install in the MFTF build)
- Use PHP 8.2
- Remove the CreateDummyCartInsightData patch from the `patch_list` table
- Run setup:upgrade

Related work items: #221783