Merged PR 47133: Fetch up to 5000 lists in admin
## What's being changed

In this PR we introduced a do/while loop to fetch all the lists in the backend configurations even if those are more than 1000.

## Why it's being changed

Currently our API fetches up to 1000 lists in a single call and we should be able to list all the lists even if those are more than a 1000 (up to 5000).

## How to review / test this change

- Go to Configuration > Customer preferences and confirm that you get all your lists in `Lists to Show` multiselect field
- Go to sync settings > list mapping and confirm that you get all the available lists in the dropdowns
- Additionally you should reduce the number of list that you get in case you don't have more than a 1000 lists in your account by reducing the int 1000 to sth lower (e.g 10) in the affected changes (see diff) in ListFetcher and Client classes

Related work items: #221827