Merged PR 47446: Improve messaging in set up integration modal
## What's being changed

The text in the 'Are you sure?' modal presented to merchants when they click 'Set Up Integration' in the Dotdigital account admin.

## Why it's being changed

The wording is confusing and suggests that migrated data in the `email_` tables might be reset as part of the process (which is incorrect). The revised text clarifies what is actually going to be reset.

## How to review / test this change

- Click 'Set Up Integration'
- Check the new text

Related work items: #206742