Merged PR 47437: Remove object manager from backend models
## What's being changed

Two backend models and some unit test code.

## Why it's being changed

For some strange legacy reason (in the backend models) we were using Object Manager to fetch a new instance of SerializerInterface, if the one injected was false (?). This has been removed.

In the tests, we had a newer alternative to mocking a collection that needed to contain a specific model mock.

## How to review / test this change

- Clear cache
- Check Configuration > Image Types, confirm the expected value lists are displayed in the selects
- Check Data Mapping > Custom Attributes. Confirm that the tool works as expected.
- Check Automation > Order Status Automation Enrolment. Confirm that events can be mapped to programs.
- See green unit test build.

Related work items: #221645