Merged PR 47697: 2.3-legacy backport for 8209d8d - Restore connection attribute in queue definitions
This PR backports a specific bugfix to the 2.3-legacy branch.

## What's being changed

Our queue XML files.

## Why it's being changed

We are rolling back a change to queue_topology.xml that we made in response to a suggestion in a Github issue. We wanted to make our queue as flexible as possible i.e. allowing merchants to use either "db" or "amqp" for it as they preferred. Unfortunately, this change breaks `setup:upgrade` in Magento < 2.4.5. To resolve both this and the original error reported in the Github issue, queue_consumer.xml now specifies a "db" connection as well.

## How to review / test this change

- Test in any version < 2.4.5
- Turn on the crons
- Run setup:upgrade (see no XML error)
- Save account credentials
- Confirm a job is processed by the sync queue

Related work items: #224163

Related work items: #224163