Merged PR 47728: Check for $contactFromTable before calling getContactId()
## What's being changed

We're adding extra checks for `$contactFromTable` before calling `getContactId()` on it.

## Why it's being changed

`$contactFromTable` can be null. Even though the contact _should_ be there following registration, it is sometimes not. For example:
- Dotdigital connection is not yet activated
- customer registers (CustomerLogin observer does nothing)
- merchant switches on Dotdigital connection
- customer refreshes their account
- `Error: call to a member function getContactId() on bool`

This scenario was exposed by an MFTF test that created a customer first, then connected to DD.

## How to review / test this change

- Run:
```
vendor/bin/mftf run:test ApiEnabledAndAtleastOnePrefEnabledAndConsentProvidedRedirectToPrefCenterTest
```
- Test should pass
- Check preferences, lists and data fields all display as expected in the account

Related work items: #224858