Merged PR 47414: Refactor avoidable usages of protected access for properties and methods
## What's being changed

A number of places where we have used "protected" property and method access. Most of the changes are simple, i.e. we didn't need to use `protected`. In some cases, though (for example the Importer sync code, and the code for Setup Integration), we have refactored the parent abstract class to minimise usage of `protected` (so that sub-classes have the specific dependencies they need injected).

## Why it's being changed

Adobe has encouraged composition over inheritance for years; the changes in this PR are prompted by an audit of our code prior to inclusion in their new Assurance Program. We've been able to remove many usages of `protected` but not all (without much more extensive refactoring). We've probably over-used inheritance in the past, although there are some fairly new systems in our code (for example data migration) where it still feels like the most suitable approach.

## How to review / test this change

- Setup Integration
- Data migration
- Importer (all sync types)
- Automation sync
- Batch processors
- System message monitors
- Any 'recommended' EDC endpoint

Related work items: #221643