Merged PR 48053: Rest execute can return string too
## What's being changed

This PR updates the return type from the `Rest::execute()` function.

## Why it's being changed

The CPAAS client we've built in the SMS module extends the core Rest class - and it can return a string response in the case of certain errors (for example, hitting https://api-cpaas.dotdigital.com/cpaas/messages with an unknown message id). This change resolves a PHPStan error in the SMS module.

## How to review / test this change

- Ensure SMS module passes php stan

Related work items: #226477