Merged PR 48309: Specify table in order status query
## What's being changed

A query in SalesDataManager.

## Why it's being changed

We had a `WHERE` clause in our query that filtered orders for certain statuses. We didn't specify the table before because natively, Magento's `sales_order_item` table has no `status` column. However it must be the case that some merchants do have such a column added, and in this case we get an exception thrown:
```
Integrity constraint violation: 1052 Column 'status' in where clause is ambiguous
```

## How to review / test this change

- Place an order as a customer
- In Dotdigital > Configuration > Data Fields observe the statuses selected for "Data fields calculation with status"
- Run customer sync
- Ensure that first and last order ids are retrieved as expected according to the statuses selected (you can put breakpoints at lines 117 and 124 of SalesDataManager to observe this)

Related work items: #227644