Merged PR 48332: Add extra check to filter out irrelevant window.postMessage calls
## What's being changed

Our engagementCloudTrial.js script. This script is responsible for listening for an `eventMessage` coming from the trial signup microsite (loaded inside of an iframe in the Magento admin).

## Why it's being changed

The script was picking up on other types of `eventMessage` posted from the microsite, in particular one from the Intercom integration (identified by a type of `intercom-snippet__ready`). These messages need filtering out so we don't see confusing / alarming 401s logged in the console.

## How to review / test this change

- Clear your account creds in Magento admin
- Launch the trial microsite
- Set up a trial to completion
- Don't see a 401 error (add a breakpoint in L16 of engagementCloudTrial.js to see what else comes in)

Related work items: #227765