Merged PR 48978: Prevent use of same lists for customer, subscriber and guest syncs
### What's being changed

1. Removed `Registry::class` and registered `ListFetcher::class` as a singleton to store records in memory.
2. Added a new Unique lists `BeforeSave()` validation.

### Why it's being changed

We removed the `Registry::class` since it is deprecated and we do not need lists to persist on save in memory for runtime. We are preventing users from pushing contacts to the same lists because it's important to ensure that subscribers are synced to their own, separate list for marketing. Merchants are advised not to mix up contacts (some to be marketed to, others not) in the same list. [Learn more](https://support.dotdigital.com/en/articles/8199589-map-lists-in-magento-open-source-and-adobe-commerce).

### How to review / test this change

1. **ListFetcher instance**: Add a breakpoint to `/Model/Lists/ListFetcher.php@L30` and observe the instance only fetching lists from dotdigital once.
2. **List mapping**: Navigate to Configuration > Dotdigital > Sync settings > List mapping.
3. **Unique lists error**: Try saving the configuration with the same lists selected in the drop-downs, and observe the error `Choose different lists to avoid mixing your Customers, Subscribers, and Guests. Learn more.`
4. Test using multi store configurations with config inheritance

### Notes

- Don't forget to run `bin/magento setup:upgrade`.
- please test with https://dev.azure.com/dotdigital/ec/_git/ec-core-magento2-extension-sms/pullrequest/49035

Related work items: #228780