Merged PR 49089: Specify tables in all order status queries
## What's being changed

Queries in SalesDataManager.

## Why it's being changed

Follow up to #da6732435fadf3c96dd4a6fedbdfa2a6fe1cc505.

We discovered an additional WHERE clause in a query that filtered orders for certain statuses, which was again breaking with:
```
Integrity constraint violation: 1052 Column 'status' in where clause is ambiguous
```

So, we're patching that one too AND setting the table alias everywhere we possibly can even in cases that could not possibly be ambiguous.

## How to review / test this change

- Start on develop
- Find a customer in your preferred local Magento that has placed several orders, ideally on different days of the week, different months, and for different products
- In Dotdigital > Configuration > Data Fields observe the statuses selected for "Data fields calculation with status"
- Run customer sync
- Now switch to this branch, reset your contact and re-run customer sync
- Compare the customer export. Do all the sales data values match?
- Pay special attention to the columns MOST_PUR_CATEGORY, MOST_FREQ_PUR_DAY, MOST_FREQ_PUR_MON
- Add a `status` column in `sales_order_item`, re-run it, see if anything else breaks

## Notes

Again, this only breaks installations that have a rogue `status` column in `sales_order_item`.

Related work items: #227644