Merged PR 49739: Add a new 'Created Date' property to be synced with catalog sync
A new 'Created Date' property has been added to the catalog sync functionality. This property will be synced along with the other catalog data during the sync process.

The 'Created Date' property provides valuable information about the creation date of each catalog item. Syncing this property allows this information to be available for further processing and analysis, aiding in various business decisions and operations.

- Update or create new records as needed in prep for a sync.
- Execute the catalog sync process.
- Verify the synced data to ensure that the 'Created Date' property is correctly synced to Dotdigital.

Please note that all catalogs seem to have a default value of `01 January 0001 00:00` for the date, so don't be confused when checking the value.

Related work items: #217093, #234451