Merged PR 50068: Restore the AC for non-subscribers loophole
## What's being changed

This PR restores some special logic that allows abandoned cart automations for contacts who are not subscribers, if:
- Sync Settings > Allow non-subscribed contacts to be imported = No
- Configuration > Allow abandoned cart for non-subscribed contacts = Yes

To achieve this I've used the imperfect (but actually very effective) inheritance model of the AbandonedCart subclass overriding select bits of the core AutomationProcessor.

## Why it's being changed

This is a loophole that allows abandoned carts to function regardless of subscriber status and restrictions on the syncing of subscribers. It's one we intentionally closed when we refactored automation sync in 4.20.0, but we're now re-opening it because a) merchants have asked for it and b) we still allow it for AC series, so we're not being consistent.

## How to review / test this change

The web of conditions governing enrolment is complex and there are a number of cases to test (see separate doc). However the basic routine is:

- Set up abandoned cart automation
- Set 'Allow non-subscribed contacts to be imported' to No
- Set 'Allow abandoned cart for non-subscribed contacts' to Yes
- Drop a cart on the storefront and identify your quote via easy email capture at checkout
- When the automation sync runs, your email should be enrolled
- Your DD contact should not be added to any list
- Confirm that if 'Allow non-subscribed contacts to be imported' is No and 'Allow abandoned cart for non-subscribed contacts' is also No, we see the same failure with the error "Non-subscribed contacts cannot be enrolled."
- Repeat with a new customer automation (do not subscribe)
- Your customer should not be enrolled (you should still see the error "Non-subscribed contacts cannot be enrolled.")

Related work items: #234986