Merged PR 50056: Remove checkCampaignEnrolmentActive method
## What's being changed

The automation sync Sender class.

## Why it's being changed

For many years we have included a pre-send check to confirm that the target program is active before sending enrolments. However this is not necessary, because the enrolment request itself returns a response that will indicate if the program was not active.

## How to review / test this change

- Set up automation enrolment in your Magento
- Once you have configured Magento, trigger an automation row
- Now set the target program to Draft in DD
- Run automation sync
- Confirm that the automation is still marked as 'Deactivated' with the expected message
- Switch the program back to Active and repeat the test. This time you should see a successful enrolment.

Related work items: #236204