Merged PR 50436: Send cart phase updates after order placed using message queue
## What's being changed

Cart phase updates are now sent using message queues, instead of via the importer.

## Why it's being changed

When an order is placed, if the order's quote id was previously associated with an abandoned cart automation, we tell Dotdigital that the quote is now ORDER_COMPLETE. Ideally this happens as soon as possible after the order is placed.

## How to review / test this change

- Turn on cron
- Set up abandoned cart program enrolment
- Drop a cart and wait for the automation to be added to the table (it doesn't matter what status)
- Complete your order
- Observe that a cart phase message is queued
- Confirm in Dotdigital that the relevant cart is updated with cart phase ORDER_COMPLETE

## Notes

All the logic for checking whether a quote id was enrolled previously is moved to the consumer to remove extra database queries from the web process.

Related work items: #235822