Merged PR 50391: Create cron to check and update PendingOptIn contacts
## What's being changed

Previously we were checking for contacts that "pending opt in" during abandoned cart (for abandoned cart program enrolments) and automation status crons in order to update their status to Confirmed, Expired or just to update the update date if their status changed.
Now we removed this logic from these crons and we created a new separate cron to handle these updates.

## Why it's being changed

Since we are abandoning automation_status cron and we are moving in to queues we need to keep this functionality alive.

## How to review / test this change
- In dotdigital platform select Enforce double opt-in for all signups, including bulk uploads in Settings > General > Account Settings
![image.png](https://dev.azure.com/dotdigital/b2ba4bc1-4a9a-48f7-a69a-5cf84134b7c4/_apis/git/repositories/cc22fc8a-89ce-4edf-ba9c-81a01b55e11a/pullRequests/50391/attachments/image.png)
- Create a few automations and confirm that status moves to PendingOptIn
- Run bin/magento dotdigital:task and select PendingContactChecker

Check based on the expire date
By default expire date is set to 24hours

a) If Contact is not opted in
If created_at date  > now - 24hours then the updated_at date should be updated to the current time
If created_at date < now - 24 hours then enrolment status should have been updated to "Expired"

b) If Contact has been opted in
Status should have been updated to Confirmed (after the task/cron run)

Same test should be confirmed for abandoned carts and the same changes should be apply in the email_abandoned_cart table

Confirm that we publish automations with "Confirmed" status to the queue.

Related work items: #236376